Various changes to math and utility libraries.
Both libraries are now const-correct.
std::string is passed by const-reference (which should be replaced by
std::string_view whenever MSVC decides to implement that >.>).
Replaced for-loops using iterators, with range-based loops (e.g. for
(auto& it : data); guarenteed to give best speed).
Changed the underlying data structure for the vector classes to use an
std::array instead of a c-style array. This allows for free
bounds-checking and super fast data manipulation, and much fewer
for-loops.
Optimized a few functions to use the STL so code is more concise and
speedy.
Replaced unsigned int with std::size_t (it basically means the same
thing but less characters to type).
Replaced abs() with the overloaded std::abs().